Improved setup user experience and added frame delay for Qml apps
Added the "Configure" dialog for setting up a valid Kraken device orientation before  connecting to the device and showing the controller interface.
Added bucket output to monitor example (hides under the FPS when it is being drawn to the display)
Renamed some files for consistency
Cleaned up some code in the device controller (krakenzdriver class), improved stability of moving from a Qml App back to a static image.
Added the ability to control frame delay for a Qml app (6-7 fps seems to be the sweet spot)
Cleaned up the logic controlling device not found and device permission issues.